requirements.txt
absl-py==0.9.0
astor==0.8.1
click==7.1.2
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python==4.2.0.34
opt-einsum==3.2.1
protobuf==3.12.2
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
